retranslator/requirements.txt
pytest-xprocess==0.12.1
pytest==5.2.2
